pyproject.toml
[project]
name = "text2sql-agentic-pipeline"
version = "1.0.0"
description = "Multi-agent LangGraph text-to-SQL pipeline"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.1.0,<2.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "tenacity>=9.0.0,<10.0.0",
    "pyyaml>=6.0.0,<7.0.0",
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "streamlit>=1.30.0,<2.0.0",
    "groq>=0.4.0,<2.0.0",
    "openai>=1.0.0,<2.0.0",
    "langchain-groq>=1.0.0",
    "langchain-openai>=0.3.0",
    "langchain-ollama>=1.0.1",
    "plotly>=5.0.0,<7.0.0",
    "ragas>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0,<10.0.0",
    "pytest-mock>=3.14.0,<4.0.0",
    "ruff>=0.8.0,<1.0.0",
]
mysql = ["mysql-connector-python>=9.0.0,<10.0.0"]
postgresql = ["psycopg2-binary>=2.9.0,<3.0.0"]
vector = [
    "pinecone>=5.1.0",
    "sentence-transformers>=5.0.0,<6.0.0",
    "chromadb>=0.5.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "database*", "graph*", "llm*", "vector*", "evaluation*", "memory*", "utils*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"
